move (new) API infra into infra-api module
