+++ avs-api/pom.xml
@@ -29,6 +29,11 @@
             <artifactId>data</artifactId>
             <version>${revision}</version>
         </dependency>
+        <dependency>
+            <groupId>org.example.age</groupId>
+            <artifactId>infra-api</artifactId>
+            <version>${revision}</version>
+        </dependency>
 
         <dependency>
             <groupId>com.google.dagger</groupId>
