+++ avs-api/src/main/java/module-info.java
@@ -8,5 +8,6 @@ module org.example.age.avs.api {
     requires org.example.age.api;
     requires org.example.age.common.api;
     requires org.example.age.data;
+    requires org.example.age.infra.api;
     requires undertow.core;
 }
